+++ plugin/src/main/java/com/stratio/cassandra/lucene/IndexServiceWide.java
@@ -18,6 +18,7 @@
 
 package com.stratio.cassandra.lucene;
 
+import com.stratio.cassandra.lucene.cache.SearchCacheUpdater;
 import com.stratio.cassandra.lucene.column.Columns;
 import com.stratio.cassandra.lucene.index.DocumentIterator;
 import com.stratio.cassandra.lucene.key.KeyMapper;
@@ -194,8 +195,9 @@ public class IndexServiceWide extends IndexService {
     @Override
     public IndexReaderWide indexReader(DocumentIterator documents,
                                        ReadCommand command,
-                                       ReadOrderGroup orderGroup) {
-        return new IndexReaderWide(command, table, orderGroup, documents, this);
+                                       ReadOrderGroup orderGroup,
+                                       SearchCacheUpdater cacheUpdater) {
+        return new IndexReaderWide(this, command, table, orderGroup, documents, cacheUpdater);
     }
 
 }