+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/util/CassandraUtilsBuilder.java
@@ -82,8 +82,8 @@ public class CassandraUtilsBuilder {
         return this;
     }
 
-    public CassandraUtilsBuilder withStaticColumn(String name, String type,boolean createMapper) {
-        columns.put(name, type+" static");
+    public CassandraUtilsBuilder withStaticColumn(String name, String type, boolean createMapper) {
+        columns.put(name, type + " static");
         if (createMapper) {
             String baseType = type.replaceAll("(.*)(<|,)", "").replace(">", "");
             SingleColumnMapper<?> mapper = defaultMapper(baseType);
@@ -94,7 +94,6 @@ public class CassandraUtilsBuilder {
         }
         return this;
     }
-
 
     public CassandraUtilsBuilder withUDT(String column, String field, String type) {
         Map<String, String> udt = udts.get(column);