+++ testsAT/src/test/java/com/stratio/cassandra/lucene/testsAT/util/CassandraUtilsSelect.java
@@ -30,7 +30,10 @@ import com.stratio.cassandra.lucene.builder.search.condition.Condition;
 import com.stratio.cassandra.lucene.builder.search.sort.SortField;
 
 import java.lang.reflect.Array;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.LinkedList;
+import java.util.List;
 
 import static com.stratio.cassandra.lucene.testsAT.util.CassandraConfig.FETCH;
 import static com.stratio.cassandra.lucene.testsAT.util.CassandraConfig.LIMIT;
@@ -196,11 +199,11 @@ public class CassandraUtilsSelect {
     public <T> CassandraUtils check(String column, Class<T> clazz, T... expecteds) {
         List<Row> rows = get();
         List<T> values = new ArrayList<>();
-        for (Row row:rows) {
+        for (Row row : rows) {
             T value = row.get(column, clazz);
             values.add(value);
         }
-        T [] actuals = (T[]) Array.newInstance(clazz, values.size());
+        T[] actuals = (T[]) Array.newInstance(clazz, values.size());
         values.toArray(actuals);
         assertArrayEquals("Expected different values", expecteds, actuals);
         return parent;